Move removeFromCart to the merlins connector
