+++ web/app/containers/cart/partials/cart-remove-item.jsx
@@ -3,7 +3,7 @@ import {connect} from 'react-redux'
 import {createPropsSelector} from 'reselect-immutable-helpers'
 import {getAssetUrl} from 'progressive-web-sdk/dist/asset-utils'
 import {CART_REMOVE_ITEM_MODAL} from '../constants'
-import {removeFromCart} from '../../../store/cart/actions'
+import {removeItem} from '../actions'
 import {closeModal} from 'progressive-web-sdk/dist/store/modals/actions'
 import {isModalOpen} from 'progressive-web-sdk/dist/store/modals/selectors'
 import {getRemoveItemID} from '../selectors'
@@ -12,7 +12,7 @@ import Sheet from 'progressive-web-sdk/dist/components/sheet'
 import Button from 'progressive-web-sdk/dist/components/button'
 import Image from 'progressive-web-sdk/dist/components/image'
 
-const CartRemoveItemModal = ({closeModal, isOpen, removeItemID, removeFromCart}) => {
+const CartRemoveItemModal = ({closeModal, isOpen, removeItemID, removeItem}) => {
     return (
         <Sheet
             className="pw--no-shadow t-cart__remove-item-confirmation-modal"
@@ -53,7 +53,7 @@ const CartRemoveItemModal = ({closeModal, isOpen, removeItemID, removeFromCart})
                         className="c--secondary u-text-uppercase u-flex u-margin-start"
                         onClick={() => {
                             closeModal()
-                            removeFromCart(removeItemID)
+                            removeItem(removeItemID)
                         }}
                     >
                         Ok
@@ -76,7 +76,7 @@ CartRemoveItemModal.propTypes = {
     /**
     * Removes the item from the cart
     */
-    removeFromCart: React.PropTypes.func,
+    removeItem: React.PropTypes.func,
     /**
     * The id of the item being deleted
     */
@@ -90,6 +90,6 @@ const mapStateToProps = createPropsSelector({
 
 const mapDispatchToProps = {
     closeModal: () => closeModal(CART_REMOVE_ITEM_MODAL),
-    removeFromCart
+    removeItem
 }
 export default connect(mapStateToProps, mapDispatchToProps)(CartRemoveItemModal)
