+++ web/app/integration-manager/cart/commands.js
@@ -7,3 +7,5 @@ export const register = (commands) => {
 export const getCart = (...args) => connector.getCart(...args)
 
 export const fetchCartPageData = (...args) => connector.fetchCartPageData(...args)
+
+export const removeFromCart = (...args) => connector.removeFromCart(...args)
